Fixed: responses in english and lowerCase - Combine controller contact and contacts, eslint
