+++ controllers/testimonies.js
@@ -1,7 +1,8 @@
 const { success, error, serverError } = require('../helpers/requestResponses');
 
 const {
-  allTestimonies, createTestimony, findTestimonyByPk, findTestimony, updateTestimonies, destroyTestimony,
+  allTestimonies, createTestimony, findTestimonyByPk, findTestimony, updateTestimonies,
+  destroyTestimony,
 } = require('../services/testimony');
 
 const getAllTestimonies = async (req, res) => {
@@ -9,7 +10,7 @@ const getAllTestimonies = async (req, res) => {
     const data = await allTestimonies();
 
     if (data) success({ res, message: 'list of all testimonies', data });
-    else error({ res, message: 'tesmimonies not found' });
+    else error({ res, message: 'testimonies not found' });
   } catch (err) {
     serverError({ res, message: err.message });
   }
@@ -21,11 +22,11 @@ const createATestimony = async (req, res) => {
   try {
     const newTestimony = await createTestimony(name, content);
 
-    if (!newTestimony) return error({ res, message: 'Testimony already exists', status: 400 });
+    if (!newTestimony) return error({ res, message: 'testimony already exists', status: 400 });
 
     return success({
       res,
-      message: 'Testimony created',
+      message: 'testimony created',
       status: 201,
       data: newTestimony,
     });
